fix(AF-968): address review findings
- gate the user-id filter on a whole UUID (a partial id was a guaranteed 400 per keystroke)
- render a dash, not a role caption, for an identity with no role
- guard slice() against (long) overflow on an unpaged page > 0
- move the website 'Privileged access' paragraph after the unused-grant knobs it was splitting
- wording nits in docs/04, docs/05, use-cases, the e2e helper comment; coverage include for the api module

+++ backend/src/main/java/com/bablsoft/accessflow/access/internal/DefaultPrivilegedAccessService.java
@@ -174,7 +174,7 @@ class DefaultPrivilegedAccessService implements PrivilegedAccessService {
         // against is the unpaged "everything" case over an empty result set.
         int size = pageRequest == null ? Math.max(rows.size(), 1) : pageRequest.size();
         int from = Math.min((int) Math.min((long) page * size, Integer.MAX_VALUE), rows.size());
-        int to = Math.min(from + size, rows.size());
+        int to = (int) Math.min((long) from + size, rows.size());
         int totalPages = (int) Math.ceil(rows.size() / (double) size);
         return new PageResponse<>(rows.subList(from, to), page, size, rows.size(), totalPages);
     }
